+++ libopenmpt/libopenmpt.h
@@ -1200,6 +1200,15 @@ LIBOPENMPT_API LIBOPENMPT_DEPRECATED int32_t openmpt_module_get_current_tempo( o
  * \since 0.7.0
  */
 LIBOPENMPT_API double openmpt_module_get_current_tempo2( openmpt_module * mod );
+/*! \brief Get the current sequence
+ *
+ * \param mod The module handle to work on.
+ * \return The sequence (order list) in which the currently playing subsong is located.
+ * \remarks There is no corresponding setter function. The current sequence is adjusted automatically by selecting a different subsong.
+ * \remarks One sequence can contain multiple subsongs. Most module formats only support a single sequence, hence the return value of this function will often be 0 even in modules that contain multiple subsongs.
+ * \since 0.9.0
+ */
+LIBOPENMPT_API int32_t openmpt_module_get_current_sequence(openmpt_module* mod);
 /*! \brief Get the current order
  *
  * \param mod The module handle to work on.