+++ libopenmpt/libopenmpt_impl.hpp
@@ -229,6 +229,7 @@ public:
 	std::int32_t get_current_speed() const;
 	std::int32_t get_current_tempo() const;
 	double get_current_tempo2() const;
+	std::int32_t get_current_sequence() const;
 	std::int32_t get_current_order() const;
 	std::int32_t get_current_pattern() const;
 	std::int32_t get_current_row() const;