+++ doc/libopenmpt/changelog.md
@@ -7,6 +7,11 @@ is just a high-level summary.
 
 ### libopenmpt 0.9.0-pre
 
+ *  [**New**] libopenmpt: New APIs for retrieving the sequence (order list) to
+    which the currently playing subsong belongs:
+    `openmpt::module::get_current sequence()` (C++) and
+    `openmpt_module_get_current_sequence()` (C). 
+
  *  libopenmpt now always uses architecture-specific SIMD intrinsics when
     available. This currently only enables the SSE2 implementation for Reverb.